Re-add sample application evolution files.
It appears that most of them no longer matter, as they have a comment
that says to enable Evolutions if you desire to use them. They are still
useful illustrations for the samples.

For play-slick-iteratees-sample, however, the evolution is still
necessary in order to create the table. And this is why the test was
failing!