Support for Slick 3.0 with Play 2.4. Fix #231
* This commit effectively drops support for older versions of Play and Slick.
This means that only Slick 3 with Play 2.4 is supported.

* Slick 3 is fundamentally different from previous versions, which is why a
major refactoring was needed. This refactoring lead to what I hope are major
simplifications, and the module's job now boils down to hook Slick into the
Play lifecycle, making sure that databases are properly closed when a Play
application is stopped or reloaded.

* Ad-hoc support for Slick with Play iteratees is no longer
needed because Slick allows to obtain a (reactive) stream when executing a
query in the db. A reactive stream can be easily converted into a enumeratee
via one of the converters available in the `play.api.libs.streams.Streams`.
See the `iteratee-sample` for an example.

* The existing magic for auto-detecting the proper Slick driver from the JDBC
driver was removed. While I understand this was a handy feature, it was also
not ideal because 1) play-slick configuration did not use to be valid Slick
configuration, and 2) it was only useful for accessing the driver for the
`default` database. Therefore, the feature was removed and users are now
requested to provide a valid Slick configuration.

* If your application.conf contains a correct Slick configuration, then you
can retrieve a Slick `DatabaseConfig` instance, holding a Slick driver and
database, either by using DI or by calling `DatabaseConfigProvider.get`.

* Added a reference.conf that can be used to provide different defaults.

* play-slick module disables both the `play.api.db.DBModule` and
`play.api.db.HikariCPModule`. The DBModule is disabled because we need to
provide a different implementation of `play.api.db.DBApi` to support Play
evolutions. While the HikariCPModule is disabled because Slick is in control
of creating (and managing) the connection pool.

* Dropped DDL support. The feature may have been convenient, but the
implementation was fragile.

* Updated user documentation, together with a migration guide.

* All samples projects have been updated and some of them are used in the
user documentation to showcase different usages of play-slick.

* build.sbt was also refactored.

* Bumped Play Slick version to 1.0!

Using Slick 3 in Play now really boils down to actually use Play and Slick
itself, with no further abstractions provided by play-slick.