Improved database evolutions support. Fix #245
After having split play-jdbc in different modules, the Play Slick module can
depend on play-jdbc-api alone, allowing it to bind `DBApi` to a different
implementation than the one provided by `play-jdbc`. The clear benefit of this
change is that we no longer need to disable modules in the reference.conf,
because both the `DBModule` and `HikariCPModule` are no longer in the classpath.

Furthermore, evolutions support was moved out of the Play Slick core, so that it
is now an opt-in feature.

Finally, bumped Play version to the minimum required version that includes the
`play-jdbc` modularisation changes I made in the upstream project. The Play
version should be changed again once Play 2.4.0-RC3 is out, and before releasing
play-slick 1.0.0-RC2.