Working around connection timeout causing spurious failures. Re #298
The PR validator was spuriously fail while runnin the `computer-database` tests.
The failure was due to a connection timeout which we already experienced in the
`iteratee-sample` project and was fixed by increasing the hikaricp connection
timeout (see commit b682ab7). Because that seems to have worked, this commits
increases the hikaricp connection timeout when running the tests of any of the
sample projects.

By the way, there is a fairly long discussion about the same problem in
brettwooldridge/HikariCP#104, but it's not clear if there is indeed a bug on the
hikaricp side, or connection are being leaked by the database connection library
(in this case Slick).