Correctly reading the database's url from a datasource's config. Fix #324
In play-slick 1.0.x we were using the connection's metadata to retrieve the
database's url. In play-slick 1.1.x we have tried to use the Slick database's
config, but the implementation didn't consider datasource's config, as the
stacktrace in the related ticket demonstrates.

The fix simply consisted in reintroducing the former, play-slick 1.0.x,
behavior, as it seemed to have been working just fine so far.