Upgraded to Play 2.5.0-RC1
I deleted a number of sample apps because a non DI approach to
play-slick is now wrong, and some of the sample apps were just too
similar to one another, from a DI perspective.  The computer-database
sample is disabled, it's selenium tests are failing, and it seems to be
a bug in htmlunit, possibly to do with submitting forms using GET.  The
same test passes with no problems using firefox.